+++ src/Ocelot.Provider.Kubernetes/IEndPointClient.cs
@@ -1,0 +1,9 @@
+﻿using KubeClient.Models;
+using KubeClient.ResourceClients;
+
+namespace Ocelot.Provider.Kubernetes;
+
+public interface IEndPointClient : IKubeResourceClient
+{
+    Task<EndpointsV1> GetAsync(string serviceName, string kubeNamespace = null, CancellationToken cancellationToken = default);
+}
